Some customization
Changed the font to “Helvetica Condensed Bold” Allows for easier
compiles on OSX. Also tweaked sizing.

Added a new object type called “Ginkgo Object”. Initializes a label
button with an additional text field so that you can have a “type”
specifier for representing different types of objects. In my case, it
was to differentiate between controllable “devices”, “layers”, “cues”,
“presets”, etc.

Increased resolution of controller labels and made increment finer.

turned off mouseUP event trigger on buttons, since this just seemed to
result in a dup event.

Created a autosize method for canvases that only autosizes vertically.

Added a function to test for mouseOver on labels. Hoping to eventually
use this to create a fine control for label values and callback.

increased slider label/value resolution. added a setValue variant
called “setMyValue” that seems to play more nicely with external
control inputs that map from -1 to 1. The old function seems to prefer
0 to 1.